# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Event Ticketing API Documentation",
        "description": "Creation of an API in PHP Laravel for event ticketing",
        "contact": {
            "email": "ronsard2014@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/events/{event}/ticket-types": {
            "get": {
                "tags": [
                    "Ticket Types"
                ],
                "summary": "Get all ticket types for a specific event with pagination",
                "description": "Retrieve the list of ticket types available for a given event, with pagination support.",
                "operationId": "getTicketTypesForEvent",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "ID of the event",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            },
            "post": {
                "tags": [
                    "Ticket Types"
                ],
                "summary": "Create a new ticket type for a specific event",
                "description": "Create and store a new ticket type associated with a specific event.",
                "operationId": "createTicketType",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "ID of the event",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ticket_type_name",
                                    "ticket_type_price",
                                    "ticket_type_quantity",
                                    "ticket_type_real_quantity",
                                    "ticket_type_total_quantity"
                                ],
                                "properties": {
                                    "ticket_type_name": {
                                        "type": "string",
                                        "example": "VIP"
                                    },
                                    "ticket_type_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 99.99
                                    },
                                    "ticket_type_quantity": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "ticket_type_real_quantity": {
                                        "type": "integer",
                                        "example": 90
                                    },
                                    "ticket_type_total_quantity": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "ticket_type_description": {
                                        "type": "string",
                                        "example": "VIP access with exclusive benefits."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ticket type created successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/v1/events/{event}/ticket-types/{ticketType}": {
            "get": {
                "tags": [
                    "Ticket Types"
                ],
                "summary": "Get a specific ticket type for a specific event",
                "description": "Retrieve details of a specific ticket type for a given event.",
                "operationId": "getTicketType",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "ID of the event",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "ticketType",
                        "in": "path",
                        "description": "ID of the ticket type",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Ticket type not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Ticket Types"
                ],
                "summary": "Update a specific ticket type for a specific event",
                "description": "Update details of a specific ticket type for a given event.",
                "operationId": "updateTicketType",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "ID of the event",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "ticketType",
                        "in": "path",
                        "description": "ID of the ticket type",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ticket_type_name",
                                    "ticket_type_price",
                                    "ticket_type_quantity",
                                    "ticket_type_real_quantity",
                                    "ticket_type_total_quantity"
                                ],
                                "properties": {
                                    "ticket_type_name": {
                                        "type": "string",
                                        "example": "VIP"
                                    },
                                    "ticket_type_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 99.99
                                    },
                                    "ticket_type_quantity": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "ticket_type_real_quantity": {
                                        "type": "integer",
                                        "example": 90
                                    },
                                    "ticket_type_total_quantity": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "ticket_type_description": {
                                        "type": "string",
                                        "example": "VIP access with exclusive benefits."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ticket type updated successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Ticket Types"
                ],
                "summary": "Delete a specific ticket type for a specific event",
                "description": "Delete a specific ticket type associated with a given event.",
                "operationId": "deleteTicketType",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "ID of the event",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "ticketType",
                        "in": "path",
                        "description": "ID of the ticket type",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Ticket type deleted successfully"
                    },
                    "404": {
                        "description": "Ticket type not found"
                    }
                }
            }
        }
    },
...
}